Add user interface and refactor API routes; integrate Header component
