+++ src/app/api/user/interface.ts
@@ -1,0 +1,3 @@
+export interface RequestBody {
+    username: string
+}
