+++ src/app/layout.js
@@ -1,5 +1,6 @@
 import localFont from 'next/font/local'
 import './globals.css'
+import Header from '@/components/Header'
 
 const geistSans = localFont({
   src: './fonts/GeistVF.woff',
@@ -63,6 +64,7 @@ export default function RootLayout({ children }) {
   return (
     <html lang='en'>
       <body className={`${geistSans.variable} ${geistMono.variable} antialiased`}>
+        <Header />
         {children}
       </body>
     </html>
